requirements.txt
docker
aiohttp
asyncio
pyyaml 
requests 
pydantic
